requirements.txt
# SOLAR PhotoSync v1.0 - Dependencies
# Python 3.9+

# Web server
aiohttp>=3.8.0

# Telegram Bot API (опционально, для расширений)
python-telegram-bot>=20.0

# Image processing
Pillow>=9.0.0
pillow-heif>=0.10.0

# Utilities
python-dateutil>=2.8.0

# For development
pytest>=7.0.0
pytest-asyncio>=0.21.0
